refactor: SourceMusic 위치 수정(#68)
