+++ src/main/java/play/pluv/login/application/LoginService.java
@@ -3,7 +3,7 @@ package play.pluv.login.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import play.pluv.member.domain.Member;
-import play.pluv.music.domain.MusicStreaming;
+import play.pluv.playlist.domain.MusicStreaming;
 import play.pluv.oauth.application.SocialLoginClientComposite;
 import play.pluv.oauth.domain.OAuthMemberInfo;
