+++ src/main/java/play/pluv/music/application/MusicExplorerComposite.java
@@ -11,9 +11,9 @@ import java.util.Set;
 import org.springframework.stereotype.Component;
 import play.pluv.music.domain.DestinationMusic;
 import play.pluv.music.domain.MusicId;
-import play.pluv.music.domain.MusicStreaming;
 import play.pluv.music.domain.SourceMusic;
 import play.pluv.music.exception.MusicException;
+import play.pluv.playlist.domain.MusicStreaming;
 
 @Component
 public class MusicExplorerComposite {
