+++ src/main/java/play/pluv/music/controller/MusicController.java
@@ -1,7 +1,7 @@
 package play.pluv.music.controller;
 
-import static play.pluv.music.domain.MusicStreaming.SPOTIFY;
-import static play.pluv.music.domain.MusicStreaming.YOUTUBE;
+import static play.pluv.playlist.domain.MusicStreaming.SPOTIFY;
+import static play.pluv.playlist.domain.MusicStreaming.YOUTUBE;
 
 import java.util.List;
 import org.springframework.web.bind.annotation.PostMapping;
