+++ src/main/java/play/pluv/music/domain/MusicId.java
@@ -1,5 +1,7 @@
 package play.pluv.music.domain;
 
+import play.pluv.playlist.domain.MusicStreaming;
+
 public record MusicId(
     MusicStreaming musicStreaming,
     String id
