+++ src/main/java/play/pluv/oauth/application/SocialLoginClientComposite.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import org.springframework.stereotype.Component;
-import play.pluv.music.domain.MusicStreaming;
+import play.pluv.playlist.domain.MusicStreaming;
 import play.pluv.oauth.domain.OAuthMemberInfo;
 import play.pluv.oauth.exception.OAuthException;
