+++ src/main/java/play/pluv/oauth/google/GoogleConnector.java
@@ -1,6 +1,6 @@
 package play.pluv.oauth.google;
 
-import static play.pluv.music.domain.MusicStreaming.YOUTUBE;
+import static play.pluv.playlist.domain.MusicStreaming.YOUTUBE;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,7 +13,7 @@ import org.springframework.util.MultiValueMap;
 import play.pluv.music.application.MusicExplorer;
 import play.pluv.music.domain.DestinationMusic;
 import play.pluv.music.domain.MusicId;
-import play.pluv.music.domain.MusicStreaming;
+import play.pluv.playlist.domain.MusicStreaming;
 import play.pluv.music.domain.SourceMusic;
 import play.pluv.oauth.application.SocialLoginClient;
 import play.pluv.oauth.domain.OAuthMemberInfo;
