+++ src/main/java/play/pluv/playlist/application/PlayListConnector.java
@@ -1,7 +1,7 @@
 package play.pluv.playlist.application;
 
 import java.util.List;
-import play.pluv.music.domain.MusicStreaming;
+import play.pluv.playlist.domain.MusicStreaming;
 import play.pluv.playlist.domain.PlayList;
 import play.pluv.playlist.domain.PlayListId;
 import play.pluv.playlist.domain.PlayListMusic;
