+++ src/main/java/play/pluv/playlist/controller/PlayListController.java
@@ -1,7 +1,7 @@
 package play.pluv.playlist.controller;
 
-import static play.pluv.music.domain.MusicStreaming.SPOTIFY;
-import static play.pluv.music.domain.MusicStreaming.YOUTUBE;
+import static play.pluv.playlist.domain.MusicStreaming.SPOTIFY;
+import static play.pluv.playlist.domain.MusicStreaming.YOUTUBE;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
