+++ src/main/java/play/pluv/playlist/domain/PlayListId.java
@@ -1,7 +1,5 @@
 package play.pluv.playlist.domain;
 
-import play.pluv.music.domain.MusicStreaming;
-
 public record PlayListId(
     String id,
     MusicStreaming musicStreaming
