+++ src/test/java/play/pluv/api/LoginApiTest.java
@@ -13,8 +13,8 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWit
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static play.pluv.music.domain.MusicStreaming.SPOTIFY;
-import static play.pluv.music.domain.MusicStreaming.YOUTUBE;
+import static play.pluv.playlist.domain.MusicStreaming.SPOTIFY;
+import static play.pluv.playlist.domain.MusicStreaming.YOUTUBE;
 
 import org.junit.jupiter.api.Test;
 import play.pluv.login.application.dto.GoogleLoginRequest;
