+++ src/test/java/play/pluv/login/application/LoginServiceTest.java
@@ -1,7 +1,7 @@
 package play.pluv.login.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static play.pluv.music.domain.MusicStreaming.SPOTIFY;
+import static play.pluv.playlist.domain.MusicStreaming.SPOTIFY;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
