+++ src/test/java/play/pluv/login/application/RegisterUpdaterTest.java
@@ -1,9 +1,8 @@
 package play.pluv.login.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static play.pluv.music.domain.MusicStreaming.SPOTIFY;
+import static play.pluv.playlist.domain.MusicStreaming.SPOTIFY;
 
-import com.jayway.jsonpath.internal.Utils;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import play.pluv.member.domain.Member;
